api/requirements/local.txt
-r base.txt

django-extensions==3.1.5
Werkzeug==2.0.3

black==22.3.0
flake8==4.0.1
mypy==0.950
isort==5.10.1

celery-stubs==0.1.2
django-stubs==1.9.0
djangorestframework-stubs==1.4.0
types-requests==2.27.25
factory-boy==3.2.1

pytest-freezegun==0.4.2
pytest-mock==3.7.0
vcrpy==4.1.1
